+++ tests/tests/test-09-mirt.model.R
@@ -23,6 +23,9 @@ test_that('syntax', {
     model7 <- mirt.model('F1 = 1-2
                          F2 = 3-5
                          START = (2, a2, 1.5), (4,a1,-1)')
+    model8 <- mirt.model('F1 = 1-2
+                         F2 = 3-5
+                         CONSTRAIN = (1, 3, a1, a2), (5, 2, a2, a1), (1-3, d)')
 
     set.seed(1234)
     group <- sample(c('male', 'female'), 1000, TRUE)
@@ -34,6 +37,7 @@ test_that('syntax', {
     mod4 <- multipleGroup(data, model4, group=group, verbose = FALSE)
     mod5 <- multipleGroup(data, model5, group=group, verbose = FALSE)
     mod7 <- mirt(data, model7, verbose=FALSE, calcNull=FALSE)
+    mod8 <- mirt(data, model8, verbose=FALSE, calcNull=FALSE)
 
     expect_equal(mod2values(mod1)$value, c(0.987973787231699, 1.85608912732841, 0, 1, 1.08103954211169, 0.808007534786952, 0, 1, 1.70595475896956, 1.80426768080187, 0, 1, 0.765076394253259, 0.486005938565521, 0, 1, 0.735771996169788, 1.85448564531374, 0, 1, 0, 1),
                  tolerance = 1e-2)
@@ -48,6 +52,8 @@ test_that('syntax', {
     expect_equal(mod2values(mod6)$value, c(1.0903077932694, 0, 1.91140861863278, 0, 1, 1.0903077932694, 0, 0.810348254879057, 0, 1, 0, 1.0101813883046, 1.46075728200956, 0, 1, 0, 1.0101813883046, 0.521336672537102, 0, 1, 0, 1.0101813883046, 1.99270549803447, 0, 1, 0, 0, 1, 0.911175950866077, 1),
                  tolerance = 1e-2)
     expect_equal(as.numeric(coef(mod7, simplify=TRUE, digits = 7)$items), c(-1.153815,-0.2728293,0,-1,0,0,1.5,1.740508,0.5660805,0.5774055,1.945502,0.9276134,1.818513,0.5439836,1.789067,0,0,0,0,0,1,1,1,1,1),
+                 tolerance = 1e-2)
+    expect_equal(as.numeric(coef(mod8, simplify=TRUE, digits = 7)$items), c(0.5501291,3.146379,0,0,0,0,0,0.5501291,0.4096282,3.146379,1.46666,1.46666,1.46666,0.4535099,3.685736,0,0,0,0,0,1,1,1,1,1),
                  tolerance = 1e-2)
 
     data(data.read, package = 'sirt')